Add a fake data generator
Since the real data has its own licence, and I'm not allowed to distribute the data,
I created a fake data generator to mimic the real data. This way the project on github
can be run without breaking.

I will/have included instructions to get the real data in the project's README.